refactor: centralize dynamic activity type naming
