+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from './jsx/createVoltraComponent.js'
+export * from './dynamic-live-activity.js'
 export * from './payload.js'
 export * from './payload/short-names.js'
 export * from './renderer/index.js'
