+++ packages/expo-plugin/jest.config.js
@@ -3,6 +3,10 @@ module.exports = {
   testEnvironment: 'node',
   testMatch: ['<rootDir>/src/**/*.node.test.ts'],
   modulePathIgnorePatterns: ['<rootDir>/build'],
+  moduleNameMapper: {
+    '^(\\.{1,2}/.*)\\.js$': '$1',
+    '^@use-voltra/core$': '<rootDir>/../core/src/index.ts',
+  },
   transform: {
     '^.+\\.tsx?$': [
       'ts-jest',
