+++ packages/expo-plugin/package.json
@@ -27,7 +27,8 @@
     "typecheck": "tsc -p tsconfig.typecheck.json --noEmit"
   },
   "dependencies": {
-    "@babel/core": "^7.27.4"
+    "@babel/core": "^7.27.4",
+    "@use-voltra/core": "workspace:^"
   },
   "keywords": [
     "voltra",
