+++ packages/expo-plugin/src/dynamic-live-activity.ts
@@ -1,13 +1,1 @@
-/**
- * Returns the generated ActivityKit attributes type name for a definition ID.
- * @experimental
- */
-export function getDynamicLiveActivityAttributesType(definitionId: string): string {
-  const upperCamelCaseId = definitionId
-    .split('_')
-    .filter(Boolean)
-    .map((segment) => `${segment[0].toUpperCase()}${segment.slice(1)}`)
-    .join('')
-
-  return `Voltra${upperCamelCaseId}LiveActivityAttributes`
-}
+export { getDynamicLiveActivityAttributesType } from '@use-voltra/core'
