+++ packages/ios-client/expo-plugin/jest.config.js
@@ -6,6 +6,7 @@ module.exports = {
   moduleNameMapper: {
     '^(\\.{1,2}/.*)\\.js$': '$1',
     '^@use-voltra/compiler$': '<rootDir>/../../compiler/src/index.ts',
+    '^@use-voltra/core$': '<rootDir>/../../core/src/index.ts',
     '^@use-voltra/expo-plugin$': '<rootDir>/../../expo-plugin/src/index.ts',
     '^@use-voltra/expo-plugin/(.*)$': '<rootDir>/../../expo-plugin/src/$1',
     '^@use-voltra/metro/scanner$': '<rootDir>/../../metro/src/scanner.ts',
