+++ packages/ios/src/live-activity/dynamic.ts
@@ -21,16 +21,4 @@ export interface DynamicLiveActivityContentState {
   props: DynamicLiveActivityProps
 }
 
-/**
- * Returns the generated ActivityKit attributes type name for a definition ID.
- * @experimental
- */
-export function getDynamicLiveActivityAttributesType(definitionId: string): string {
-  const upperCamelCaseId = definitionId
-    .split('_')
-    .filter(Boolean)
-    .map((segment) => `${segment[0].toUpperCase()}${segment.slice(1)}`)
-    .join('')
-
-  return `Voltra${upperCamelCaseId}LiveActivityAttributes`
-}
+export { getDynamicLiveActivityAttributesType } from '@use-voltra/core'
